+++ infrastructure-detekt/src/main/kotlin/DetektPlugin.kt
@@ -49,7 +49,7 @@ private fun Project.configureDependencies() {
     }
 
     dependencies {
-        detektPlugins("io.gitlab.arturbosch.detekt:detekt-formatting:1.21.0")
+        detektPlugins("io.gitlab.arturbosch.detekt:detekt-formatting:1.22.0")
     }
 }